gpu.h: move SelectView into gpu.h
